refactor(desktop): extract the update store into updater/store and shared updater/support
Mechanical move out of updater.ts: the store layout/ownership machinery
(sentinel, owned-root validation, manual-clear rebuild, owned subdirs),
the persisted store metadata types with their guards, and the store path
constants move to updater/store.ts. Generic helpers they depend on
(createError, JSON guards, path containment, atomic JSON IO) move to
updater/support.ts for the remaining split slices to share. updater.ts
imports the same surface back; no exports change and no behavior
changes.
