+++ apps/desktop/src/main/updater/support.ts
@@ -1,0 +1,59 @@
+import { mkdir, readdir, readFile, rename, writeFile } from "node:fs/promises";
+import { dirname, isAbsolute, relative } from "node:path";
+
+import type { DesktopUpdateErrorSnapshot } from "@open-design/sidecar-proto";
+
+/**
+ * @module updater/support
+ *
+ * Generic helpers shared across the updater modules: structured error
+ * construction, JSON type-guard primitives, path containment, and atomic
+ * JSON file IO. No updater state and no store semantics.
+ */
+
+export function createError(code: string, message: string, details?: unknown): DesktopUpdateErrorSnapshot {
+  return {
+    code,
+    ...(details === undefined ? {} : { details }),
+    message,
+  };
+}
+
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value != null && !Array.isArray(value);
+}
+
+export function stringField(record: Record<string, unknown>, key: string): string | null {
+  const value = record[key];
+  return typeof value === "string" && value.length > 0 ? value : null;
+}
+
+export function containsPath(root: string, path: string): boolean {
+  const rel = relative(root, path);
+  return rel === "" || (rel.length > 0 && !rel.startsWith("..") && !isAbsolute(rel));
+}
+
+export async function writeJson(path: string, payload: unknown): Promise<void> {
+  await mkdir(dirname(path), { recursive: true });
+  const tmp = `${path}.${process.pid}.${Date.now()}.tmp`;
+  await writeFile(tmp, `${JSON.stringify(payload, null, 2)}\n`, "utf8");
+  await rename(tmp, path);
+}
+
+export async function readJson<T>(path: string): Promise<T | null> {
+  try {
+    return JSON.parse(await readFile(path, "utf8")) as T;
+  } catch {
+    return null;
+  }
+}
+
+export async function readJsonStrict<T>(path: string): Promise<T> {
+  return JSON.parse(await readFile(path, "utf8")) as T;
+}
+
+export async function directoryIsEmpty(path: string): Promise<boolean> {
+  const entries = await readdir(path);
+  return entries.length === 0;
+}
+
